+++ apps/kyb-app/src/components/organisms/UIRenderer/elements/JSONForm/json-form.fields.ts
@@ -3,16 +3,18 @@ import { CountryPicker } from '@/components/organisms/UIRenderer/elements/JSONFo
 import { DocumentField } from '@/components/organisms/UIRenderer/elements/JSONForm/components/DocumentField';
 import { IndustriesPicker } from '@/components/organisms/UIRenderer/elements/JSONForm/components/IndustriesPicker';
 import { JSONFormArrayFieldLayout } from '@/components/organisms/UIRenderer/elements/JSONForm/components/JSONFormArrayFieldLayout';
+import { FieldTemplate } from '@/components/organisms/UIRenderer/elements/JSONForm/components/FieldTemplate';
 import { LocalePicker } from '@/components/organisms/UIRenderer/elements/JSONForm/components/LocalePicker';
 import { Multiselect } from '@/components/organisms/UIRenderer/elements/JSONForm/components/Multiselect/Multiselect';
 import { NationalityPicker } from '@/components/organisms/UIRenderer/elements/JSONForm/components/NationalityPicker';
+import { RelationshipDropdown } from '@/components/organisms/UIRenderer/elements/JSONForm/components/RelationshipDropdown';
 import { StatePicker } from '@/components/organisms/UIRenderer/elements/JSONForm/components/StatePicker';
 import { withDynamicUIInput } from '@/components/organisms/UIRenderer/elements/JSONForm/hocs/withDynamicUIInput';
 import {
   AutocompleteTextInputAdapter,
   baseLayouts,
   BooleanFieldAdapter,
-  DateInputAdater,
+  DateInputAdapter,
   FileInputAdapter,
   PhoneInputAdapter,
   TextInputAdapter,
@@ -25,7 +27,7 @@ export const jsonFormFields = {
 
   // Component with suffix Input is an extend of supported field types
   FileInput: withDynamicUIInput(FileInputAdapter),
-  DateInput: withDynamicUIInput(DateInputAdater),
+  DateInput: withDynamicUIInput(DateInputAdapter),
   PhoneInput: withDynamicUIInput(PhoneInputAdapter),
   AutocompleteInput: withDynamicUIInput(AutocompleteTextInputAdapter),
   DocumentInput: withDynamicUIInput(DocumentField),
@@ -36,9 +38,11 @@ export const jsonFormFields = {
   IndustriesPicker: withDynamicUIInput(IndustriesPicker),
   Multiselect: withDynamicUIInput(Multiselect),
   StatePicker: withDynamicUIInput(StatePicker),
+  RelationshipDropdown: withDynamicUIInput(RelationshipDropdown),
 };
 
 export const jsonFormLayouts = {
   ...baseLayouts,
+  FieldTemplate,
   ArrayFieldTemplate: JSONFormArrayFieldLayout,
 };