+++ packages/ui/src/components/molecules/inputs/DatePickerInput/DatePickerInput.tsx
@@ -1,11 +1,11 @@
 import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
 import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
 import { DatePicker } from '@mui/x-date-pickers/DatePicker';
-import { FocusEvent, useCallback, useMemo, useState } from 'react';
-import { CalendarDays, ChevronLeft, ChevronRight } from 'lucide-react';
+import { FocusEvent, FunctionComponent, useCallback, useMemo, useState } from 'react';
 import dayjs, { Dayjs } from 'dayjs';
 import { TextField, TextFieldProps, ThemeProvider } from '@mui/material';
 import { muiTheme } from '@/common/mui-theme';
+import { CalendarDays, ChevronLeft, ChevronRight } from 'lucide-react';
 import { Paper } from '@/components/atoms';
 
 export interface DatePickerChangeEvent {
@@ -17,10 +17,16 @@ export interface DatePickerChangeEvent {
 
 export type DatePickerValue = number | string | Date | null;
 
+export interface DatePickerParams {
+  disableFuture?: boolean;
+  disablePast?: boolean;
+}
+
 export interface DatePickerProps {
   value?: DatePickerValue;
   name?: string;
   disabled?: boolean;
+  params?: DatePickerParams;
   onChange: (event: DatePickerChangeEvent) => void;
   onBlur?: (event: FocusEvent<any>) => void;
 }
@@ -29,6 +35,7 @@ export const DatePickerInput = ({
   value: _value,
   name,
   disabled = false,
+  params,
   onChange,
   onBlur,
 }: DatePickerProps) => {
@@ -46,12 +53,19 @@ export const DatePickerInput = ({
     (value: Dayjs | null) => {
       if (!value) return onChange({ target: { value: null, name } });
 
-      onChange({
-        target: {
-          value: serializeValue(value),
-          name,
-        },
-      });
+      try {
+        const serializedDateValue = serializeValue(value);
+        onChange({
+          target: {
+            value: serializedDateValue,
+            name,
+          },
+        });
+      } catch (error) {
+        // Ignoring serialization due to partial date input
+        // Attept to serialize partially entered date e.g 12/MM/YYYY will cause exception
+        return null;
+      }
     },
     [name, onChange, serializeValue],
   );
@@ -62,8 +76,8 @@ export const DatePickerInput = ({
     return deserializeValue(_value);
   }, [_value, deserializeValue]);
 
-  const Field = useMemo(
-    () => (props: TextFieldProps) => {
+  const Field = useMemo(() => {
+    const Component: FunctionComponent<TextFieldProps> = props => {
       return (
         <TextField
           {...props}
@@ -100,14 +114,16 @@ export const DatePickerInput = ({
           }}
         />
       );
-    },
-    [isFocused, onBlur],
-  );
+    };
+
+    return Component;
+  }, [isFocused]);
 
   return (
     <ThemeProvider theme={muiTheme}>
       <LocalizationProvider dateAdapter={AdapterDayjs}>
         <DatePicker
+          {...params}
           disabled={disabled}
           value={value}
           onChange={handleChange}