+++ packages/workflow-core/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @ballerine/workflow-core
 
+## 0.5.33
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.34
+
+## 0.5.32
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.33
+
+## 0.5.31
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.32
+
+## 0.5.30
+
+### Patch Changes
+
+- added helper for transformers
+
 ## 0.5.29
 
 ### Patch Changes