+++ sdks/workflow-browser-sdk/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @ballerine/workflow-browser-sdk
 
+## 0.5.33
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.34
+  - @ballerine/workflow-core@0.5.33
+
+## 0.5.32
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.33
+  - @ballerine/workflow-core@0.5.32
+
+## 0.5.31
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.32
+  - @ballerine/workflow-core@0.5.31
+
+## 0.5.30
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.5.30
+
 ## 0.5.29
 
 ### Patch Changes